+++ Fonts.lua
@@ -1,8 +1,6 @@
 local _, addon = ...
 local LSM = LibStub("LibSharedMedia-3.0")
 
--- Define custom font files located in \NiceDamage\fonts\
--- Add fonts here if you want to add custom fonts ["Font Name"] = "fontfile.ttf"
 local myFonts = {
     ["Pepsi Modern"] = "pepsi_modern.ttf",
     ["Pepsi Cursive"] = "pepsi_cursive.ttf",
@@ -18,8 +16,6 @@ local myFonts = {
 
 function addon:RegisterFonts()
     local prefix = "Interface\\AddOns\\NiceDamage\\fonts\\"
-    
-    -- Loop through the table and register each entry with LSM
     for name, file in pairs(myFonts) do
         LSM:Register("font", name, prefix .. file)
     end